requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pandas==2.0.3
openai==0.28.1
Werkzeug==2.3.7
gunicorn==21.2.0